controle-financeiro/requirements.txt
python-telegram-bot
reportlab
python-dotenv
supabase
psycopg2-binary
Flask 